// api/src/container/Product/types.ts
export enum ROUTES {
  createCategory = "createCategory",
  getAllCategory = "getAllCategory",
  createSubCategory = "createSubCategory",
  getSubCategoryByCategoryId = "getSubCategoryByCategoryId",
  createProduct = "createProduct",
  getProductBySubCategoryId = "getProductBySubCategoryId",
  updateProductImage = "updateProductImage",
  searchProduct = "searchProduct",
  similarProduct = "similarProduct",
  getProduct = "getProduct",
}